moved auth store location and removed unused variables
